+++ src/template_new_wasm_mod.rs
@@ -67,12 +67,12 @@ fn copy_to_files(rust_project_name: &str, github_owner_or_organization: &str, we
     let url = "https://github.com/automation-tasks-rs/cargo_auto_template_new_wasm/releases/latest/download/template.tar.gz";
     let reqwest_client = reqwest::blocking::Client::new();
     let http_response = reqwest_client.get(url).send();
-    if http_response.is_err() {
-        panic!("Error while retrieving data: {:#?}", http_response.err());
-    } else {
-        let body = http_response.unwrap().bytes().unwrap();
+    if let Ok(body) = http_response {
+        let body = body.bytes().unwrap();
         // Get the content of the response
         std::fs::write(path, &body).unwrap_or_else(|_| panic!("Download failed for {file_name}"));
+    } else {
+        panic!("Error while retrieving data: {:#?}", http_response.err());
     }
 
     // decompress into folder_path
@@ -100,9 +100,16 @@ fn copy_to_files(rust_project_name: &str, github_owner_or_organization: &str, we
     let mut traverse_reverse: Vec<walkdir::DirEntry> = walkdir::WalkDir::new(folder_path).into_iter().filter_map(Result::ok).collect();
     traverse_reverse.reverse();
     for entry in traverse_reverse.iter() {
-        if entry.file_name() == "cargo_auto_template_new_wasm" {
+        if entry.file_name().to_string_lossy().contains("cargo_auto_template_new_wasm") {
             println!("rename: {}", entry.path().to_string_lossy());
-            std::fs::rename(entry.path(), entry.path().parent().unwrap().join(rust_project_name)).unwrap();
+            std::fs::rename(
+                entry.path(),
+                entry
+                    .path()
+                    .to_string_lossy()
+                    .replace("cargo_auto_template_new_wasm", rust_project_name),
+            )
+            .unwrap();
         }
     }
 }